pyproject.toml
[project]
name = "llm-compliance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.6",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
